Implemented and fixed bugs to allow for playing and pausing WAV file. Created project that shows implementation of engine/backend code. Other cleanup of code.

+++ BackBeat/src/Platform/Windows/WindowsInput.cpp
@@ -27,7 +27,7 @@ namespace BackBeat {
 		double x, y;
 		glfwGetCursorPos(window, &x, &y);
 
-		return MousePos(x, y); }
+		return MousePos((float)x, (float)y); }
 
 	float Input::GetMouseX() { return GetMousePosition().x; }
